refactor encode to CursorEncoder, decode to CursorDecoder
