+++ order.go
@@ -8,3 +8,10 @@ const (
 	ASC  Order = "ASC"
 	DESC Order = "DESC"
 )
+
+func flip(order Order) Order {
+	if order == ASC {
+		return DESC
+	}
+	return ASC
+}
