+++ paginator.go
@@ -1,13 +1,9 @@
 package paginator
 
 import (
-	"encoding/base64"
-	"encoding/json"
-	"errors"
 	"fmt"
 	"reflect"
 	"strings"
-	"time"
 
 	"github.com/iancoleman/strcase"
 	"github.com/jinzhu/gorm"
@@ -25,13 +21,12 @@ func New() *Paginator {
 
 // Paginator a builder doing pagination
 type Paginator struct {
-	cursor   Cursor
-	next     Cursor
-	keys     []string
-	keyKinds []kind
-	sqlKeys  []string
-	limit    int
-	order    Order
+	cursor    Cursor
+	next      Cursor
+	keys      []string
+	tableKeys []string
+	limit     int
+	order     Order
 }
 
 // SetAfterCursor sets paging after cursor
@@ -67,30 +62,14 @@ func (p *Paginator) GetNextCursor() Cursor {
 // Paginate paginates data
 func (p *Paginator) Paginate(stmt *gorm.DB, out interface{}) *gorm.DB {
 	p.initOptions()
-	p.initKeyKinds(out)
 	p.initTableKeys(stmt, out)
-	result := p.appendPagingQuery(stmt).Find(out)
+	result := p.appendPagingQuery(stmt, out).Find(out)
 	// out must be a pointer or gorm will panic above
 	elems := reflect.ValueOf(out).Elem()
 	if elems.Kind() == reflect.Slice && elems.Len() > 0 {
 		p.postProcess(out)
 	}
 	return result
-}
-
-// Encode encodes fields to cursor
-func (p *Paginator) Encode(v interface{}) string {
-	return base64.StdEncoding.EncodeToString(p.marshalJSON(v))
-}
-
-// Decode decodes cursor to fields
-func (p *Paginator) Decode(cursor string) []interface{} {
-	b, err := base64.StdEncoding.DecodeString(cursor)
-	// @TODO: return proper error
-	if err != nil {
-		return nil
-	}
-	return p.unmarshalJSON(b)
 }
 
 /* private */
@@ -107,39 +86,24 @@ func (p *Paginator) initOptions() {
 	}
 }
 
-func (p *Paginator) initKeyKinds(out interface{}) {
-	rt := reflect.ValueOf(out).Type()
-	for rt.Kind() == reflect.Slice || rt.Kind() == reflect.Ptr {
-		rt = rt.Elem()
-	}
-	if rt.Kind() != reflect.Struct {
-		// element of out must be struct, if not, just pass it to gorm to handle the error
-		return
-	}
-	p.keyKinds = make([]kind, len(p.keys))
-	for i, key := range p.keys {
-		field, _ := rt.FieldByName(key)
-		p.keyKinds[i] = toKind(field.Type)
-	}
-}
-
 func (p *Paginator) initTableKeys(db *gorm.DB, out interface{}) {
 	table := db.NewScope(out).TableName()
 	for _, key := range p.keys {
-		p.sqlKeys = append(p.sqlKeys, fmt.Sprintf("%s.%s", table, strcase.ToSnake(key)))
+		p.tableKeys = append(p.tableKeys, fmt.Sprintf("%s.%s", table, strcase.ToSnake(key)))
 	}
 }
 
-func (p *Paginator) appendPagingQuery(stmt *gorm.DB) *gorm.DB {
+func (p *Paginator) appendPagingQuery(stmt *gorm.DB, out interface{}) *gorm.DB {
+	decoder := NewCursorDecoder(out, p.keys...)
 	var fields []interface{}
 	if p.hasAfterCursor() {
-		fields = p.Decode(*p.cursor.After)
+		fields = decoder.Decode(*p.cursor.After)
 	} else if p.hasBeforeCursor() {
-		fields = p.Decode(*p.cursor.Before)
+		fields = decoder.Decode(*p.cursor.Before)
 	}
 	if len(fields) > 0 {
 		stmt = stmt.Where(
-			p.getCursorQuery(p.getOperator()),
+			p.getCursorQuery(),
 			p.getCursorQueryArgs(fields)...,
 		)
 	}
@@ -156,10 +120,11 @@ func (p *Paginator) hasBeforeCursor() bool {
 	return !p.hasAfterCursor() && p.cursor.Before != nil
 }
 
-func (p *Paginator) getCursorQuery(op string) string {
-	qs := make([]string, len(p.sqlKeys))
+func (p *Paginator) getCursorQuery() string {
+	qs := make([]string, len(p.tableKeys))
+	op := p.getOperator()
 	composite := ""
-	for i, sqlKey := range p.sqlKeys {
+	for i, sqlKey := range p.tableKeys {
 		qs[i] = fmt.Sprintf("%s%s %s ?", composite, sqlKey, op)
 		composite = fmt.Sprintf("%s%s = ? AND ", composite, sqlKey)
 	}
@@ -186,8 +151,8 @@ func (p *Paginator) getOrder() string {
 	if p.hasBeforeCursor() {
 		order = flip(p.order)
 	}
-	orders := make([]string, len(p.sqlKeys))
-	for index, sqlKey := range p.sqlKeys {
+	orders := make([]string, len(p.tableKeys))
+	for index, sqlKey := range p.tableKeys {
 		orders[index] = fmt.Sprintf("%s %s", sqlKey, order)
 	}
 	return strings.Join(orders, ", ")
@@ -202,108 +167,14 @@ func (p *Paginator) postProcess(out interface{}) {
 	if p.hasBeforeCursor() {
 		elems.Set(reverse(elems))
 	}
+	encoder := NewCursorEncoder(p.keys...)
 	if p.hasBeforeCursor() || hasMore {
-		cursor := p.Encode(elems.Index(elems.Len() - 1))
+		cursor := encoder.Encode(elems.Index(elems.Len() - 1))
 		p.next.After = &cursor
 	}
 	if p.hasAfterCursor() || (hasMore && p.hasBeforeCursor()) {
-		cursor := p.Encode(elems.Index(0))
+		cursor := encoder.Encode(elems.Index(0))
 		p.next.Before = &cursor
 	}
 	return
-}
-
-func (p *Paginator) marshalJSON(value interface{}) []byte {
-	rv, ok := value.(reflect.Value)
-	if !ok {
-		rv = reflect.ValueOf(value)
-	}
-	for rv.Kind() == reflect.Ptr {
-		rv = rv.Elem()
-	}
-	fields := make([]interface{}, len(p.keys))
-	for i, key := range p.keys {
-		fields[i] = rv.FieldByName(key).Interface()
-	}
-	// @TODO: return proper error
-	b, _ := json.Marshal(fields)
-	return b
-}
-
-func (p *Paginator) unmarshalJSON(bytes []byte) []interface{} {
-	var fields []interface{}
-	err := json.Unmarshal(bytes, &fields)
-	// @TODO: return proper error
-	if err != nil {
-		return nil
-	}
-	return p.castJSONFields(fields)
-}
-
-func (p *Paginator) castJSONFields(fields []interface{}) []interface{} {
-	result := make([]interface{}, len(fields))
-	for i, field := range fields {
-		kind := p.keyKinds[i]
-		switch f := field.(type) {
-		case bool:
-			bv, err := castJSONBool(f, kind)
-			if err != nil {
-				return nil
-			}
-			result[i] = bv
-		case float64:
-			fv, err := castJSONFloat(f, kind)
-			if err != nil {
-				return nil
-			}
-			result[i] = fv
-		case string:
-			sv, err := castJSONString(f, kind)
-			if err != nil {
-				return nil
-			}
-			result[i] = sv
-		default:
-			// invalid field
-			return nil
-		}
-	}
-	return result
-}
-
-var (
-	errInvalidFieldType = errors.New("invalid field type")
-)
-
-func castJSONBool(value bool, kind kind) (interface{}, error) {
-	if kind != kindBool {
-		return nil, errInvalidFieldType
-	}
-	return value, nil
-}
-
-func castJSONFloat(value float64, kind kind) (interface{}, error) {
-	switch kind {
-	case kindInt:
-		return int(value), nil
-	case kindUint:
-		return uint(value), nil
-	case kindFloat:
-		return value, nil
-	}
-	return nil, errInvalidFieldType
-}
-
-func castJSONString(value string, kind kind) (interface{}, error) {
-	if kind != kindString && kind != kindTime {
-		return nil, errInvalidFieldType
-	}
-	if kind == kindString {
-		return value, nil
-	}
-	tv, err := time.Parse(time.RFC3339Nano, value)
-	if err != nil {
-		return nil, errInvalidFieldType
-	}
-	return tv, nil
 }
